+++ api/simplemovement.h
@@ -3,34 +3,34 @@
 
 #include "api.h"
 
-void drive(int power);
-void strafe(int power);
+void drive(byte power);
+void strafe(byte power);
 //rotate function should be depracated, nearly uncontrolable
-void rotate(int power);
-void swingTurn(int direction, int power);
-void moveDiagonal(int vector, int power);
+void rotate(byte power);
+void swingTurn(byte direction, byte power);
+void moveDiagonal(byte vector, byte power);
 
 //Need to be implemented
-void driveTicks(int ticks, int power);
-void driveIn(int dist, int power);
-void driveCm(int dist, int power);
+void driveTicks(byte ticks, byte power);
+void driveIn(byte dist, byte power);
+void driveCm(byte dist, byte power);
 
-void strafeTicks(int ticks, int power);
-void strafeIn(int dist, int power);
-void strafeCm(int dist, int power);
+void strafeTicks(byte ticks, byte power);
+void strafeIn(byte dist, byte power);
+void strafeCm(byte dist, byte power);
 
-void rotateTicks(int ticks, int power);
-void rotateDeg(int degs, int power);
-void rotateRad(int rads, int power);
+void rotateTicks(byte ticks, byte power);
+void rotateDeg(byte degs, byte power);
+void rotateRad(byte rads, byte power);
 
 //Other ideas:
-void swingTurnTicks(int ticks, int direction, int power);
-void swingTurnDeg(int degs, int directions, int power);
-void swingTurnRad(int rads, int directions, int power);
+void swingTurnTicks(byte ticks, byte direction, byte power);
+void swingTurnDeg(byte degs, byte directions, byte power);
+void swingTurnRad(byte rads, byte directions, byte power);
 
-void moveDiagonalTicks(int dist, int vector, int power);
-void moveDiagonalIn(int dist, int vector, int power);
-void moveDiagonalCm(int dist, int vector, int power);
+void moveDiagonalTicks(byte dist, byte vector, byte power);
+void moveDiagonalIn(byte dist, byte vector, byte power);
+void moveDiagonalCm(byte dist, byte vector, byte power);
 
 //Low priority: (Daniel's ideas)
-void turnSpiral(byte power, int distance, int deg);
+void turnSpiral(byte power, byte distance, byte deg);
